+++ tesseraql-identity/src/main/resources/io/tesseraql/identity/pack/default/sql/clear-role-permissions.sql
@@ -1,0 +1,4 @@
+delete from tql_role_permissions
+where
+  role_id in (select role_id from tql_roles where role_code = /* roleCode */ 'x')
+;